一些修改

+++ Egg-main/src/main/java/edu/xiyou/andrew/Egg/net/CrawlDatum.java
@@ -34,7 +34,7 @@ public class CrawlDatum {
      */
     protected String url;
 
-    public CrawlDatum(String url, byte status, byte fetchTime) {
+    public CrawlDatum(String url, byte status, long fetchTime) {
         this.status = status;
         this.fetchTime = fetchTime;
         this.url = url;
